backend/requirements.txt
fastapi
sqlalchemy
python-jose
passlib
fastapi-mail